python 3 updates

+++ modularExponentiation.py
@@ -1,6 +1,5 @@
 # Modular Exponentiation Algorithm
 # Implemented by William M Mortl
-# Coded for Python 2.7.9
 # O(n^3)
 # python modularExponentiation.py 5 2 17
 
@@ -10,10 +9,10 @@ import sys
 
 # modular exponentiation x ^ y mod N
 def modularExponentiation(x, y, N):
-	x = long(x)
-	y = long(y)
-	N = long(N)
-	retVal = long(1)
+	x = int(x)
+	y = int(y)
+	N = int(N)
+	retVal = 1
 	if (y > 0):
 		z = modularExponentiation(x, y / 2, N)
 		retVal = z ** 2
@@ -29,8 +28,8 @@ if __name__ == "__main__":
 		print("Usage: python modularExponentiation.py {val} {to the power} {modulo}")
 		print("Example: python modularExponentiation.py 5 2 17\r\n")
 	else:
-		x = long(sys.argv[1])
-		y = long(sys.argv[2])
-		N = long(sys.argv[3])
+		x = int(sys.argv[1])
+		y = int(sys.argv[2])
+		N = int(sys.argv[3])
 		v = modularExponentiation(x, y, N)
 		print(("\r\n%s to the power %s modulo %s = %s\r\n") % (str(x), str(y), str(N), str(v)))
